+++ guide/gatsby-node.js
@@ -22,8 +22,16 @@ function addElmLoader(config) {
 }
 
 function addSvgLoaders(config) {
-  var svgoConf = require('cultureamp-style-guide/webpack/svgo.config.js');
-  var spriteLoaderConf = {
+  let svgoConf = require('cultureamp-style-guide/webpack/svgo.config.js');
+
+  // Note: the SVGO Conf has a regular expression, which can't be used with JSON.stringify.
+  // Here we replace it with a plain string, which works well enough for now.
+  let convertColorsPlugin = svgoConf.plugins.find(
+    plugin => plugin.convertColors
+  );
+  convertColorsPlugin.convertColors.currentColor = '#000000';
+
+  let spriteLoaderConf = {
     symbolId: 'ca-icon-[name]',
     spriteModule: require.resolve('cultureamp-style-guide/util/svg-sprite.js'),
   };